requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
async-timeout==5.0.1
asyncssh==2.21.1
atpublic==6.0.2
attrs==25.3.0
Authlib==1.6.4
billiard==4.2.2
blinker==1.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
cyclopts==3.24.0
databricks-sdk==0.67.0
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
dpath==2.2.0
dulwich==0.24.2
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
email-validator==2.3.0
entrypoints==0.4
exceptiongroup==1.3.0
fastapi==0.118.0
fastmcp==2.12.4
filelock==3.19.1
Flask==3.1.2
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gto==1.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
lazy-object-proxy==1.12.0
lightgbm==4.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mcp==1.15.0
mdurl==0.1.2
mlflow==3.4.0
mlflow-skinny==3.4.0
mlflow-tracing==3.4.0
more-itertools==10.8.0
multidict==6.6.4
networkx==3.4.2
numpy==2.2.6
nvidia-nccl-cu12==2.28.3
omegaconf==2.3.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parse==1.20.2
pathable==0.4.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.15.3
scmrepo==3.5.2
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
sqltrie==0.11.2
sse-starlette==3.0.2
starlette==0.48.0
tabulate==0.9.0
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.14
Werkzeug==3.1.1
xgboost==3.0.5
yarl==1.20.1
zc.lockfile==4.0
zipp==3.23.0
